fastapi_app/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.8.0
pydantic-settings>=2.4.0
email-validator>=2.2.0
SQLAlchemy[asyncio]>=2.0.30
asyncpg>=0.29.0
aiosqlite>=0.20.0
alembic>=1.13.2
python-jose[cryptography]>=3.3.0
bcrypt>=4.1.0,<6
structlog>=24.2.0
python-dotenv>=1.0.1

# tests
pytest>=8.2.0
pytest-asyncio>=0.23.7
httpx>=0.27.0
asgi-lifespan>=2.1.0
